Moved new function objects into their own files
